refactor error handling to an interface to reuse this approach for other external tasks
