+++ src/main/java/uk/gov/hmcts/reform/waworkflowapi/clients/service/idempotency/IdempotencyTaskWorkerHandler.java
@@ -6,7 +6,6 @@ import org.camunda.bpm.client.task.ExternalTask;
 import org.camunda.bpm.client.task.ExternalTaskService;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.waworkflowapi.clients.model.idempotencykey.IdempotentId;
-import uk.gov.hmcts.reform.waworkflowapi.exceptions.IdempotencyTaskWorkerException;
 
 import java.util.Optional;
 
@@ -17,15 +16,13 @@ import static uk.gov.hmcts.reform.waworkflowapi.clients.service.idempotency.Idem
 @Slf4j
 public class IdempotencyTaskWorkerHandler {
 
-    public static final int NUMBER_OF_RETRIES = 3;
-    public static final int NO_MORE_RETRIES = 1;
-    public static final int INCIDENT_SIGNAL = 0;
-
     private final IdempotencyTaskService idempotencyTaskService;
+    private final ExternalTaskErrorHandling externalTaskErrorHandlingWithThreeRetries;
 
-    public IdempotencyTaskWorkerHandler(
-        IdempotencyTaskService idempotencyTaskService) {
+    public IdempotencyTaskWorkerHandler(IdempotencyTaskService idempotencyTaskService,
+                                        ExternalTaskErrorHandling externalTaskErrorHandlingWithThreeRetries) {
         this.idempotencyTaskService = idempotencyTaskService;
+        this.externalTaskErrorHandlingWithThreeRetries = externalTaskErrorHandlingWithThreeRetries;
     }
 
     public void checkIdempotency(ExternalTask externalTask, ExternalTaskService externalTaskService) {
@@ -37,29 +34,8 @@ public class IdempotencyTaskWorkerHandler {
                 () -> completeTask(externalTask, externalTaskService)
             );
         } catch (Exception e) {
-            handleError(externalTask, externalTaskService);
+            externalTaskErrorHandlingWithThreeRetries.handleError(externalTask, externalTaskService);
         }
-    }
-
-    private void handleError(ExternalTask externalTask, ExternalTaskService externalTaskService) {
-        if (externalTask.getRetries() == null) {
-            setFailure(externalTask, externalTaskService, NUMBER_OF_RETRIES);
-        } else if (externalTask.getRetries() > NO_MORE_RETRIES) {
-            setFailure(externalTask, externalTaskService, externalTask.getRetries() - 1);
-        } else {
-            setFailure(externalTask, externalTaskService, INCIDENT_SIGNAL);
-            throw new IdempotencyTaskWorkerException("*** ERROR with idempotency worker *** ");
-        }
-    }
-
-    private void setFailure(ExternalTask externalTask, ExternalTaskService externalTaskService, int retries) {
-        externalTaskService.handleFailure(
-            externalTask.getId(),
-            "david error message",
-            "david error details",
-            retries,
-            1000
-        );
     }
 
     private void completeTask(ExternalTask externalTask, ExternalTaskService externalTaskService) {
